Significant improvements to unit tests
