+++ go.mod
@@ -6,4 +6,5 @@ require (
 	github.com/Unleash/unleash-client-go/v3 v3.1.1
 	github.com/julienschmidt/httprouter v1.3.0
 	github.com/spf13/viper v1.6.2
+	github.com/stretchr/testify v1.2.2
 )
